requirements.txt
PyPDF2
pdfminer.six
pdfplumber
pdf2image
pytesseract
numpy==1.26.4
python-dotenv
openai
dotenv